correction checker,leaks a corriger

+++ push_swap/libft/ft_atoi.c
@@ -6,17 +6,18 @@
 /*   By: hgandar <hgandar@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 15:22:24 by hgandar           #+#    #+#             */
-/*   Updated: 2024/01/17 16:43:15 by hgandar          ###   ########.fr       */
+/*   Updated: 2024/01/22 15:56:50 by hgandar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	result;
-	int	sign;
+	int		i;
+	long	result;
+	int		sign;
 
 	i = 0;
 	sign = 1;
